Updated st3play with experimental AdLib support

+++ Frameworks/modplay/modplay/dbopl.h
@@ -43,8 +43,8 @@ typedef  unsigned char		Bit8u;
 typedef    signed char		Bit8s;
 typedef unsigned short		Bit16u;
 typedef   signed short		Bit16s;
-typedef  unsigned long		Bit32u;
-typedef    signed long		Bit32s;
+typedef  unsigned int		Bit32u;
+typedef    signed int		Bit32s;
 #ifdef _MSC_VER
 typedef unsigned __int64	Bit64u;
 typedef   signed __int64	Bit64s;
